Added Timer class for events and PhysicsEngine for physics in game. Collision was added in PhysicsEngine class.

+++ src/Renderer/Sprite.h
@@ -22,7 +22,7 @@ namespace RenderEngine {
 
 		struct FrameDescription
 		{
-			FrameDescription(const glm::vec2 _leftBottomUV, glm::vec2 _rightTopUV, const uint64_t _duration)
+			FrameDescription(const glm::vec2 _leftBottomUV, glm::vec2 _rightTopUV, const double _duration)
 				: leftBottomUV(_leftBottomUV)
 				, rightTopUV(_rightTopUV)
 				, duration(_duration)
@@ -30,7 +30,7 @@ namespace RenderEngine {
 
 			glm::vec2 leftBottomUV;
 			glm::vec2 rightTopUV;
-			uint64_t duration;
+			double duration;
 
 		};
 
@@ -49,7 +49,7 @@ namespace RenderEngine {
 					const size_t frameId = 0) const;
 
 		void insertFrames(std::vector<FrameDescription> framesDescriptions);
-		uint64_t getFrameDuration(const size_t frameId) const;
+		double getFrameDuration(const size_t frameId) const;
 		size_t getFramesCount() const;
