+++ src/Resources/ResourceManager.cpp
@@ -272,7 +272,11 @@ bool ResourceManager::loadJSONResources(const std::string& JSONPath)
 			{
 				continue;
 			}
-
+			//// for tests
+			//if (name == "shield")
+			//{
+			//	int a = 1;
+			//}
 			auto framesIt = currentSprite.FindMember("frames");
 			if (framesIt != currentSprite.MemberEnd())
 			{
@@ -282,7 +286,7 @@ bool ResourceManager::loadJSONResources(const std::string& JSONPath)
 				for (const auto& currentFrame : framesArray)
 				{
 					const std::string subTextureStr = currentFrame["subTexture"].GetString();
-					const uint64_t duration = currentFrame["duration"].GetUint64();
+					const double duration = currentFrame["duration"].GetDouble();
 					const auto pTextureAtlas = getTexture(textureAtlas);
 					const auto pSubTexture = pTextureAtlas->getSubTexture(subTextureStr);
 					frameDescroptions.emplace_back(pSubTexture.leftBottomUV, pSubTexture.rightTopUV, duration);
